refactor and organize ipc files
